Move draw functions into folder.
