+++ react-ui/src/components/table/draw/mappedPositions.js
