+++ react-ui/src/components/table/draw/player.js
