+++ react-ui/src/components/table/draw/playerCard.js
